Remove SolvableFrameSystem
