+++ motionplan/constraint_test.go
@@ -82,8 +82,6 @@ func TestConstraintPath(t *testing.T) {
 }
 
 func TestLineFollow(t *testing.T) {
-	logger := golog.NewDebugLogger("armplay")
-
 	p1 := spatial.NewPoseFromProtobuf(&commonpb.Pose{
 		X:  440,
 		Y:  -447,
@@ -139,20 +137,19 @@ func TestLineFollow(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 	err = fs.AddFrame(markerFrame, m)
 	test.That(t, err, test.ShouldBeNil)
-	fss := NewSolvableFrameSystem(fs, logger)
 
 	solveFrame := markerFrame
 	goalFrame := fs.World()
 
-	sFrames, err := fss.TracebackFrame(solveFrame)
+	sFrames, err := fs.TracebackFrame(solveFrame)
 	test.That(t, err, test.ShouldBeNil)
 
 	// Create a frame to solve for, and an IK solver with that frame.
 	sf, err := newSolverFrame(
-		fss,
+		fs,
 		sFrames,
 		goalFrame.Name(),
-		frame.StartPositions(fss),
+		frame.StartPositions(fs),
 	)
 	test.That(t, err, test.ShouldBeNil)
