+++ motionplan/motionPlanner_test.go
@@ -2,6 +2,7 @@ package motionplan
 
 import (
 	"context"
+	"math"
 	"math/rand"
 	"strconv"
 	"testing"
@@ -125,11 +126,19 @@ func TestPlanningWithGripper(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 	err = fs.AddFrame(gripper, ur5e)
 	test.That(t, err, test.ShouldBeNil)
-	fss := NewSolvableFrameSystem(fs, logger.Sugar())
-	zeroPos := frame.StartPositions(fss)
+	zeroPos := frame.StartPositions(fs)
 
 	newPose := frame.NewPoseInFrame("gripper", spatialmath.NewPoseFromPoint(r3.Vector{100, 100, 0}))
-	solutionMap, err := fss.SolvePose(context.Background(), zeroPos, newPose, gripper.Name())
+	solutionMap, err := PlanMotion(
+		context.Background(),
+		logger.Sugar(),
+		newPose,
+		gripper,
+		zeroPos,
+		fs,
+		nil,
+		nil,
+	)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, len(solutionMap), test.ShouldBeGreaterThanOrEqualTo, 2)
 }
@@ -249,4 +258,197 @@ func testPlanner(t *testing.T, plannerFunc seededPlannerConstructor, config plan
 		}, cfg.Options.Resolution)
 		test.That(t, ok, test.ShouldBeTrue)
 	}
+}
+
+
+func makeTestFS(t *testing.T) frame.FrameSystem {
+	t.Helper()
+	fs := frame.NewEmptySimpleFrameSystem("test")
+
+	urOffset, err := frame.NewStaticFrame("urOffset", spatialmath.NewPoseFromPoint(r3.Vector{100, 100, 200}))
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(urOffset, fs.World())
+	gantryOffset, err := frame.NewStaticFrame("gantryOffset", spatialmath.NewPoseFromPoint(r3.Vector{-50, -50, -200}))
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(gantryOffset, fs.World())
+
+	gantryX, err := frame.NewTranslationalFrame("gantryX", r3.Vector{1, 0, 0}, frame.Limit{math.Inf(-1), math.Inf(1)})
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(gantryX, gantryOffset)
+	gantryY, err := frame.NewTranslationalFrame("gantryY", r3.Vector{0, 1, 0}, frame.Limit{math.Inf(-1), math.Inf(1)})
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(gantryY, gantryX)
+
+	modelXarm, err := frame.ParseModelJSONFile(utils.ResolveFile("components/arm/xarm/xarm6_kinematics.json"), "")
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(modelXarm, gantryY)
+
+	modelUR5e, err := frame.ParseModelJSONFile(utils.ResolveFile("components/arm/universalrobots/ur5e.json"), "")
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(modelUR5e, urOffset)
+
+	// Note that positive Z is always "forwards". If the position of the arm is such that it is pointing elsewhere,
+	// the resulting translation will be similarly oriented
+	urCamera, err := frame.NewStaticFrame("urCamera", spatialmath.NewPoseFromPoint(r3.Vector{0, 0, 30}))
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(urCamera, modelUR5e)
+
+	// Add static frame for the gripper
+	bc, _ := spatialmath.NewBoxCreator(r3.Vector{200, 200, 200}, spatialmath.NewPoseFromPoint(r3.Vector{Z: 100}), "")
+	xArmVgripper, err := frame.NewStaticFrameWithGeometry("xArmVgripper", spatialmath.NewPoseFromPoint(r3.Vector{Z: 200}), bc)
+	test.That(t, err, test.ShouldBeNil)
+	fs.AddFrame(xArmVgripper, modelXarm)
+
+	return fs
+}
+
+func TestFrameSystemSolver(t *testing.T) {
+	fs := makeTestFS(t)
+	positions := frame.StartPositions(fs)
+	pointXarmGripper := spatialmath.NewPoseFromPoint(r3.Vector{157., -50, -288})
+	transformPoint, err := fs.Transform(positions, frame.NewPoseInFrame("xArmVgripper", spatialmath.NewZeroPose()), frame.World)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, spatialmath.PoseAlmostCoincident(transformPoint.(*frame.PoseInFrame).Pose(), pointXarmGripper), test.ShouldBeTrue)
+
+	// Set a goal such that the gantry and arm must both be used to solve
+	goal1 := spatialmath.NewPoseFromOrientation(r3.Vector{X: 257, Y: 2100, Z: -300}, &spatialmath.OrientationVectorDegrees{OZ: -1})
+	newPos, err := PlanMotion(
+		context.Background(),
+		logger.Sugar(),
+		frame.NewPoseInFrame(frame.World, goal1),
+		fs.Frame("xArmVgripper"),
+		positions,
+		fs,
+		nil,
+		nil,
+	)
+	test.That(t, err, test.ShouldBeNil)
+	solvedPose, err := fs.Transform(newPos[len(newPos)-1], frame.NewPoseInFrame("xArmVgripper", spatialmath.NewZeroPose()), frame.World)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, spatialmath.PoseAlmostCoincidentEps(solvedPose.(*frame.PoseInFrame).Pose(), goal1, 0.01), test.ShouldBeTrue)
+
+	// Solve such that the ur5 and xArm are pointing at each other, 60mm from gripper to camera
+	goal2 := spatialmath.NewPoseFromOrientation(r3.Vector{Z: 60}, &spatialmath.OrientationVectorDegrees{OZ: -1})
+	newPos, err = PlanMotion(
+		context.Background(),
+		logger.Sugar(),
+		frame.NewPoseInFrame("urCamera", goal2),
+		fs.Frame("xArmVgripper"),
+		positions,
+		fs,
+		nil,
+		map[string]interface{}{"max_ik_solutions": 100, "timeout": 150.0},
+	)
+	test.That(t, err, test.ShouldBeNil)
+
+	// Both frames should wind up at the goal relative to one another
+	solvedPose, err = fs.Transform(newPos[len(newPos)-1], frame.NewPoseInFrame("xArmVgripper", spatialmath.NewZeroPose()), "urCamera")
+	test.That(t, err, test.ShouldBeNil)
+	solvedPose2, err := fs.Transform(newPos[len(newPos)-1], frame.NewPoseInFrame("urCamera", spatialmath.NewZeroPose()), "xArmVgripper")
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, spatialmath.PoseAlmostCoincidentEps(solvedPose.(*frame.PoseInFrame).Pose(), goal2, 0.1), test.ShouldBeTrue)
+	test.That(t, spatialmath.PoseAlmostCoincidentEps(solvedPose2.(*frame.PoseInFrame).Pose(), goal2, 0.1), test.ShouldBeTrue)
+}
+
+func TestSliceUniq(t *testing.T) {
+	fs := makeTestFS(t)
+	slice := []frame.Frame{}
+	slice = append(slice, fs.Frame("urCamera"))
+	slice = append(slice, fs.Frame("gantryOffset"))
+	slice = append(slice, fs.Frame("xArmVgripper"))
+	slice = append(slice, fs.Frame("urCamera"))
+	uniqd := uniqInPlaceSlice(slice)
+	test.That(t, len(uniqd), test.ShouldEqual, 3)
+}
+
+func TestSolverFrameGeometries(t *testing.T) {
+	fs := makeTestFS(t)
+	sFrames, err := fs.TracebackFrame(fs.Frame("xArmVgripper"))
+	test.That(t, err, test.ShouldBeNil)
+	sf, err := newSolverFrame(fs, sFrames, frame.World, frame.StartPositions(fs))
+	test.That(t, err, test.ShouldBeNil)
+
+	sfPlanner, err := newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	position, err := sfPlanner.PlanSingleWaypoint(
+		context.Background(),
+		sf.sliceToMap(make([]frame.Input, len(sf.DoF()))),
+		spatialmath.NewPoseFromPoint(r3.Vector{300, 300, 100}),
+		nil,
+		nil,
+	)
+	test.That(t, err, test.ShouldBeNil)
+	gf, _ := sf.Geometries(position[len(position)-1])
+	test.That(t, gf, test.ShouldNotBeNil)
+	gripperCenter := gf.Geometries()["xArmVgripper"].Pose().Point()
+	test.That(t, spatialmath.R3VectorAlmostEqual(gripperCenter, r3.Vector{300, 300, 0}, 1e-2), test.ShouldBeTrue)
+}
+
+func TestMovementWithGripper(t *testing.T) {
+	// TODO(rb): move these tests to a separate repo eventually, as they take up too much time for general CI pipeline
+	t.Skip()
+
+	// setup solverFrame and planning query
+	fs := makeTestFS(t)
+	fs.RemoveFrame(fs.Frame("urOffset"))
+	sFrames, err := fs.TracebackFrame(fs.Frame("xArmVgripper"))
+	test.That(t, err, test.ShouldBeNil)
+	sf, err := newSolverFrame(fs, sFrames, frame.World, frame.StartPositions(fs))
+	test.That(t, err, test.ShouldBeNil)
+	goal := spatialmath.NewPoseFromOrientation(r3.Vector{500, 0, -300}, &spatialmath.OrientationVector{OZ: -1})
+	zeroPosition := sf.sliceToMap(make([]frame.Input, len(sf.DoF())))
+
+	// linearly plan with the gripper
+	motionConfig := make(map[string]interface{})
+	motionConfig["motion_profile"] = LinearMotionProfile
+	sfPlanner, err := newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	solution, err := sfPlanner.PlanSingleWaypoint(context.Background(), zeroPosition, goal, nil, motionConfig)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, solution, test.ShouldNotBeNil)
+
+	// plan around the obstacle with the gripper
+	obstacle, err := spatialmath.NewBox(spatialmath.NewPoseFromPoint(r3.Vector{300, 0, -400}), r3.Vector{50, 500, 500}, "")
+	test.That(t, err, test.ShouldBeNil)
+	geometries := make(map[string]spatialmath.Geometry)
+	geometries["obstacle"] = obstacle
+	obstacles := []*commonpb.GeometriesInFrame{frame.GeometriesInFrameToProtobuf(frame.NewGeometriesInFrame(frame.World, geometries))}
+	worldState := &commonpb.WorldState{Obstacles: obstacles}
+	sfPlanner, err = newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	solution, err = sfPlanner.PlanSingleWaypoint(context.Background(), zeroPosition, goal, worldState, nil)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, solution, test.ShouldNotBeNil)
+
+	// plan with end of arm with gripper attached - this will fail
+	sFrames, err = fs.TracebackFrame(fs.Frame("xArm6"))
+	test.That(t, err, test.ShouldBeNil)
+	sf, err = newSolverFrame(fs, sFrames, frame.World, frame.StartPositions(fs))
+	test.That(t, err, test.ShouldBeNil)
+	goal = spatialmath.NewPoseFromOrientation(r3.Vector{500, 0, -100}, &spatialmath.OrientationVector{OZ: -1})
+	zeroPosition = sf.sliceToMap(make([]frame.Input, len(sf.DoF())))
+	sfPlanner, err = newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	_, err = sfPlanner.PlanSingleWaypoint(context.Background(), zeroPosition, goal, worldState, motionConfig)
+	test.That(t, err, test.ShouldNotBeNil)
+
+	// remove linear constraint and try again
+	sfPlanner, err = newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	solution, err = sfPlanner.PlanSingleWaypoint(context.Background(), zeroPosition, goal, worldState, nil)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, solution, test.ShouldNotBeNil)
+
+	// remove gripper and try with linear constraint
+	fs.RemoveFrame(fs.Frame("xArmVgripper"))
+	sFrames, err = fs.TracebackFrame(fs.Frame("xArm6"))
+	test.That(t, err, test.ShouldBeNil)
+	sf, err = newSolverFrame(fs, sFrames, frame.World, frame.StartPositions(fs))
+	test.That(t, err, test.ShouldBeNil)
+	zeroPosition = sf.sliceToMap(make([]frame.Input, len(sf.DoF())))
+	sfPlanner, err = newPlanManager(sf, fs, logger.Sugar(), 1)
+	test.That(t, err, test.ShouldBeNil)
+	solution, err = sfPlanner.PlanSingleWaypoint(context.Background(), zeroPosition, goal, worldState, motionConfig)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, solution, test.ShouldNotBeNil)
 }
